+++ resources.tf
@@ -14,14 +14,3 @@ resource "azurerm_app_service_plan" "plan" {
 
   tags = merge(local.default_tags, var.extra_tags)
 }
-
-module "diagnostics" {
-  source  = "claranet/diagnostic-settings/azurerm"
-  version = "4.0.2"
-
-  resource_id           = azurerm_app_service_plan.plan.id
-  logs_destinations_ids = var.logs_destinations_ids
-  log_categories        = var.logs_categories
-  metric_categories     = var.logs_metrics_categories
-  retention_days        = var.logs_retention_days
-}