+++ variables.tf
@@ -28,12 +28,6 @@ variable "resource_group_name" {
   type        = string
 }
 
-variable "name_prefix" {
-  description = "Optional prefix for the generated name"
-  type        = string
-  default     = ""
-}
-
 variable "sku" {
   description = "A sku block. See documentation https://www.terraform.io/docs/providers/azurerm/r/app_service_plan.html#sku"
   type        = map(string)
@@ -50,37 +44,8 @@ variable "extra_tags" {
   default     = {}
 }
 
-variable "custom_name" {
-  description = "Name of the App Service Plan, generated if not set."
-  type        = string
-  default     = ""
-}
-
 variable "reserved" {
   description = "Flag indicating if App Service Plan should be reserved. Forced to true if \"kind\" is \"Linux\"."
   type        = string
   default     = "false"
-}
-
-variable "logs_destinations_ids" {
-  type        = list(string)
-  description = "List of destination resources Ids for logs diagnostics destination. Can be Storage Account, Log Analytics Workspace and Event Hub. No more than one of each can be set. Empty list to disable logging."
-}
-
-variable "logs_categories" {
-  type        = list(string)
-  description = "Log categories to send to destinations."
-  default     = null
-}
-
-variable "logs_metrics_categories" {
-  type        = list(string)
-  description = "Metrics categories to send to destinations."
-  default     = null
-}
-
-variable "logs_retention_days" {
-  type        = number
-  description = "Number of days to keep logs on storage account"
-  default     = 30
 }